user profile being built, everything from store displaying properly on it, edit user sends info and receives updated user from backend, need to send to store

+++ src/components/Review.js
@@ -8,7 +8,11 @@ const Review = ({ game_name, content, rating, username, id, user_id, user, handl
 
   const handleClick = () => {
     deleteReview(id)
+    
+    if (handleDelete) {
     handleDelete(id)
+    }
+    alert("Review Deleted!")
   }
 
   return (
